fix(oob): target levels as JSON ints; deploy never removes the bridge for a missing image [MESHSAT-756]
TargetInfo.Levels was a byte slice, which encoding/json renders as
base64 ("AQID" instead of [1,2,3]) and broke the level dropdown on the
live kit. Now ints.

ci-deploy.sh: on 2026-09-02 the image pull timed out on parallax's
flapping WiFi, the script stopped and removed the running container,
and compose failed with "No such image", leaving the kit without a
bridge. The script now aborts before touching the container when the
image is not present locally after the pull.

+++ internal/oob/service_test.go
@@ -453,9 +453,9 @@ func contains(list []string, s string) bool {
 	return false
 }
 
-func contains3(list []byte, b byte) bool {
+func contains3(list []int, b byte) bool {
 	for _, v := range list {
-		if v == b {
+		if v == int(b) {
 			return true
 		}
 	}
